refactor: moved/renamed shared components
