+++ src/components/SingleProduct/SheetForm/generateColors.js
